Allow new Context to be created as not dismissable
Trying to do this triggered an SQL error: Column 'dismissable' cannot be
null.

When the Dismissable checbox is unchecked, the 'new_dismissable' GPC
param is not set, and thus defaulted to empty array, so an invalid value
is passed on to save the record.

Forcing the Context's dismissable property to be false when the GPC
param is not set resolves the problem.

Fixes #19