Fix js code to show/hide the last context warning
The initial version relied on jQuery's siblings() method, to determine
the number of contexts, but the code does not work when the page is
processing 2 or more announcements.

To calculate the number, we now go back to the first category row, then
back down until the next spacer row (or the end of the table), excluding
deleted contexts.

Fixes #31