Consistent display of default in Location selects
Currently, the plugin only handles a single Location type (header).

When creating a new announcement, the code preselects the only available
value in the Location list whereas in the Edit Announcements page, when
clicking on the + button to add a new context, the selection list
defaults to (select one) even though there is only a single value to
choose from.

This commit fixes the inconsistent behavior by moving the logic into
Announce::print_location_option_list() method, so that when there is
only a single location type available, it is always preselected.

Fixes #38